requirements.txt
runpod
loguru
pillow
boto3
requests 
python-dotenv
openai==0.28
